# 10x faster AtomicBool implementation