It's very straightforward to use _actual_ atomic instead of mutex. And it's paying off in nearly 10x margin in terms of performance
```
goos: linux
goarch: amd64
pkg: github.com/junegunn/fzf/src/util
BenchmarkMutex_Set-4            20000000                69.4 ns/op
BenchmarkMutex_Get-4            20000000                68.9 ns/op
BenchmarkAtomic_Set-4           200000000                7.66 ns/op
BenchmarkAtomic_Get-4           2000000000               0.42 ns/op
```