@brualan 2020-03-25
Also, `AtomicBool` is used only here https://github.com/junegunn/fzf/blob/master/src/matcher.go#L150 for, basically, context cancellation. I plan later to use proper methods for that task. As explained here https://blog.golang.org/context, for example