@brualan 2020-03-29
I'm think that code in latest commit is much better and shouldn't get any complains from linter. Please review